ROBOT
This is not my best work. I am not happy with the result,

I have worked on this robot since after school till the time of this submission, trying to figure things out transitioning over from a different program I considered myself good at. But it's not where I want it to be, so after I submit this I will continue to tweak and edit it, remodeling what is needed to oimporve and get better.